% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{build_line_plot}
\alias{build_line_plot}
\title{Generic line plot builder}
\usage{
build_line_plot(
  data,
  x_var,
  y_var,
  color_var,
  linetype_var = NULL,
  facet_formula,
  title,
  subtitle = "",
  y_lab,
  color_lab = "Method",
  linetype_lab = "neg:pos",
  hline = NULL,
  hline_style = "dotted",
  x_breaks = NULL,
  y_limits = NULL,
  use_aggregate_theme = FALSE,
  point_size = 1.5,
  line_width = 1
)
}
\arguments{
\item{data}{Data frame}

\item{x_var}{X variable (unquoted)}

\item{y_var}{Y variable (unquoted)}

\item{color_var}{Color variable (unquoted)}

\item{linetype_var}{Optional linetype variable (unquoted), NULL for none}

\item{facet_formula}{Facet formula (e.g., \code{q0_lab ~ q1_lab})}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{y_lab}{Y-axis label}

\item{color_lab}{Color legend label}

\item{linetype_lab}{Linetype legend label (if applicable)}

\item{hline}{Optional horizontal line y-intercept}

\item{hline_style}{"dashed" or "dotted"}

\item{x_breaks}{X-axis breaks}

\item{y_limits}{Optional y-axis limits (c(min, max))}

\item{use_aggregate_theme}{Use larger theme for aggregate plots}

\item{point_size}{Point size}

\item{line_width}{Line width}
}
\description{
Generic line plot builder
}
