% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{eif_continuous_point_and_ci}
\alias{eif_continuous_point_and_ci}
\title{EIF estimator for continuous surrogates with calibration}
\usage{
eif_continuous_point_and_ci(
  Y_cal,
  Yhat_cal,
  Yhat_test,
  calibration_method = "linear",
  alpha = 0.1
)
}
\description{
EIF estimator for continuous surrogates with calibration
}
