% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{eif_discrete_point_and_ci}
\alias{eif_discrete_point_and_ci}
\title{EIF estimator with per-category calibration (optimal for discrete)}
\usage{
eif_discrete_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
Learns g: {1,...,K} -> R such that g(z) = E\link{Y | Yhat = z}
For discrete Yhat, this is just the conditional mean per category
}
