% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{eif_gam_discrete_point_and_ci}
\alias{eif_gam_discrete_point_and_ci}
\title{EIF estimator with GAM calibration}
\usage{
eif_gam_discrete_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
Fits g(Yhat) using a GAM with smooth term
For discrete Yhat with few categories, this behaves similarly to per-category
}
