% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{eif_point_and_ci}
\alias{eif_point_and_ci}
\title{EIF (Efficient Influence Function) estimator for binary surrogates.}
\usage{
eif_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.05)
}
\arguments{
\item{Y_cal}{Vector of true labels on the calibration set.}

\item{Yhat_cal}{Vector of surrogate predictions (0/1) on the calibration set.}

\item{Yhat_test}{Vector of surrogate predictions (0/1) on the test set.}

\item{alpha}{Miscoverage level for the confidence interval.}
}
\value{
A list containing:
\describe{
\item{theta}{Point estimate of the prevalence.}
\item{var}{Estimated variance of the estimator.}
\item{ci_lower}{Lower bound of the confidence interval.}
\item{ci_upper}{Upper bound of the confidence interval.}
\item{mu0_hat}{Estimated \eqn{E[Y | \hat Y = 0]}.}
\item{mu1_hat}{Estimated \eqn{E[Y | \hat Y = 1]}.}
\item{p_hat}{Proportion of \eqn{\hat Y = 1} in test set.}
}
}
\description{
Implements the post-stratified efficient estimator from the EIF theory.
For binary surrogates, this estimates \eqn{\theta = E[Y]} using:
\deqn{\hat\theta_{EIF} = (1-\hat p)\hat\mu_0 + \hat p \hat\mu_1}
where \eqn{\hat\mu_0, \hat\mu_1} are the conditional means of Y given
\eqn{\hat Y = 0, 1} estimated from calibration data, and \eqn{\hat p}
is the proportion of \eqn{\hat Y = 1} in the test sample.
}
\details{
This estimator is asymptotically efficient in Model A (surrogates are
conditionally independent of outcomes given covariates).
}
