% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{fit_calibration_linear}
\alias{fit_calibration_linear}
\title{Fit calibration model using linear regression}
\usage{
fit_calibration_linear(Y_cal, Yhat_cal)
}
\arguments{
\item{Y_cal}{Vector of true labels on the calibration set.}

\item{Yhat_cal}{Vector of surrogate predictions on the calibration set.}
}
\value{
A function that maps new surrogate values to calibrated predictions.
}
\description{
Fit calibration model using linear regression
}
