% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{fit_calibration_spline}
\alias{fit_calibration_spline}
\title{Fit calibration model using cubic spline}
\usage{
fit_calibration_spline(Y_cal, Yhat_cal, df = 4)
}
\arguments{
\item{Y_cal}{Vector of true labels on the calibration set.}

\item{Yhat_cal}{Vector of surrogate predictions on the calibration set.}

\item{df}{Degrees of freedom for the spline basis.}
}
\value{
A function that maps new surrogate values to calibrated predictions.
}
\description{
Fit calibration model using cubic spline
}
