% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{fit_reppi_model}
\alias{fit_reppi_model}
\title{Fit a RePPI-style logistic regression for \eqn{g(S,X)}.}
\usage{
fit_reppi_model(covariates, S, Y)
}
\arguments{
\item{covariates}{Data frame of covariates on the calibration set.}

\item{S}{Calibration surrogate outputs (0/1).}

\item{Y}{Calibration human labels (0/1).}
}
\value{
Fitted glm predicting Y from S and covariates (with interactions).
}
\description{
Fit a RePPI-style logistic regression for \eqn{g(S,X)}.
}
