% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{g_confusion_inverse}
\alias{g_confusion_inverse}
\title{Confusion-matrix inverse g-transform (multiclass RG).}
\usage{
g_confusion_inverse(S, context)
}
\arguments{
\item{S}{Integer vector of surrogate labels.}

\item{context}{List containing inverse confusion matrix \code{W_inv}.}
}
\value{
Matrix whose rows are \code{W_inv \%*\% e_S}.
}
\description{
Confusion-matrix inverse g-transform (multiclass RG).
}
