% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{g_identity_multiclass}
\alias{g_identity_multiclass}
\title{Identity g-transform for multiclass surrogates.}
\usage{
g_identity_multiclass(S, context)
}
\arguments{
\item{S}{Integer vector of surrogate class labels in \eqn{\{1,...,K\}}.}

\item{context}{List containing at least \code{K}.}
}
\value{
Matrix with one-hot rows representing S.
}
\description{
Identity g-transform for multiclass surrogates.
}
