% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{g_reppi_predict}
\alias{g_reppi_predict}
\title{Predict using the RePPI-style logistic regression.}
\usage{
g_reppi_predict(S, covariates, model)
}
\arguments{
\item{S}{Surrogate outputs (0/1).}

\item{covariates}{Data frame of covariates.}

\item{model}{Fitted glm from \code{fit_reppi_model}.}
}
\value{
Predicted probabilities, \eqn{P(Y = 1 \mid S, X)}.
}
\description{
Predict using the RePPI-style logistic regression.
}
