% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{generate_discrete_dgp}
\alias{generate_discrete_dgp}
\title{Generate discrete prediction data}
\usage{
generate_discrete_dgp(N, mu = c(1, 5, 9), sigma = 1, seed = NULL)
}
\arguments{
\item{N}{Sample size}

\item{mu}{Vector of means for each mixture component}

\item{sigma}{Standard deviation (same for all components)}

\item{seed}{Random seed}
}
\value{
List with Z, Y, Yhat
}
\description{
Z ~ Unif({1, ..., K}), Y | Z ~ N(mu\link{Z}, sigma^2), Yhat = Z
}
