% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{generate_plot_suite}
\alias{generate_plot_suite}
\title{Generate all 5 standard plots for a given data subset}
\usage{
generate_plot_suite(
  ci_data,
  q_data,
  facet_formula,
  filename_suffix,
  subtitle = "",
  save_fn,
  x_breaks = NULL,
  linetype_var = NULL,
  use_aggregate = FALSE,
  width = 14,
  height = 12,
  point_size = 1.5
)
}
\arguments{
\item{ci_data}{CI summary data (filtered)}

\item{q_data}{Q bias data (filtered)}

\item{facet_formula}{Facet formula}

\item{filename_suffix}{Suffix for filenames}

\item{subtitle}{Plot subtitle}

\item{save_fn}{Function to save plots (takes plot and filename)}

\item{x_breaks}{X-axis breaks}

\item{linetype_var}{Optional linetype variable for aggregate plots}

\item{use_aggregate}{Use aggregate theme}

\item{width}{Plot width}

\item{height}{Plot height}

\item{point_size}{Point size}
}
\description{
Generate all 5 standard plots for a given data subset
}
