% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_method_color_scale}
\alias{get_method_color_scale}
\title{Get the fixed color scale for methods}
\usage{
get_method_color_scale(methods = NULL)
}
\arguments{
\item{methods}{Character vector of method names to include}
}
\value{
ggplot2 scale_color_manual
}
\description{
Get the fixed color scale for methods
}
