% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{naive_continuous_point_and_ci}
\alias{naive_continuous_point_and_ci}
\title{Naive estimator (just use model predictions)}
\usage{
naive_continuous_point_and_ci(Yhat_test, alpha = 0.1)
}
\description{
Naive estimator (just use model predictions)
}
