% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{naive_discrete_point_and_ci}
\alias{naive_discrete_point_and_ci}
\title{Naive estimator: just use mean of discrete predictions}
\usage{
naive_discrete_point_and_ci(Yhat_test, alpha = 0.1)
}
\description{
Naive estimator: just use mean of discrete predictions
}
