% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{ppi_continuous_point_and_ci}
\alias{ppi_continuous_point_and_ci}
\title{PPI estimator for continuous scores}
\usage{
ppi_continuous_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
PPI estimator for continuous scores
}
