% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{ppi_pp_bootstrap}
\alias{ppi_pp_bootstrap}
\title{Bootstrap PPI++ estimator}
\usage{
ppi_pp_bootstrap(
  Y_L,
  f_L,
  f_U,
  n_boot = 500,
  alpha = 0.05,
  lambda_range = c(0, 1)
)
}
\arguments{
\item{Y_L}{Vector of human labels on the calibration set.}

\item{f_L}{Surrogate predictions on the calibration set.}

\item{f_U}{Surrogate predictions on the unlabeled/test set.}

\item{n_boot}{Number of bootstrap replicates.}

\item{alpha}{Miscoverage level for the confidence interval.}

\item{lambda_range}{Interval to constrain lambda.}
}
\value{
List with theta, lambda, var (bootstrap), and CI endpoints.
}
\description{
Bootstrap PPI++ estimator
}
