% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{ppi_pp_continuous_point_and_ci}
\alias{ppi_pp_continuous_point_and_ci}
\title{PPI++ estimator for continuous scores}
\usage{
ppi_pp_continuous_point_and_ci(
  Y_cal,
  Yhat_cal,
  Yhat_test,
  alpha = 0.1,
  lambda_range = c(0, 1)
)
}
\arguments{
\item{Y_cal}{Vector of true labels on the calibration set.}

\item{Yhat_cal}{Vector of surrogate predictions on the calibration set.}

\item{Yhat_test}{Vector of surrogate predictions on the test set.}

\item{alpha}{Miscoverage level for the confidence interval.}

\item{lambda_range}{Numeric length-2 vector giving the optimization interval.}
}
\description{
PPI++ estimator for continuous scores
}
