% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{ppi_pp_discrete_point_and_ci}
\alias{ppi_pp_discrete_point_and_ci}
\title{PPI++ estimator for discrete predictions}
\usage{
ppi_pp_discrete_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
Uses closed-form optimal lambda (no constraints):
lambda* = Cov(Y,Yhat) / \link{Var(Yhat_cal) + (m/n) * Var(Yhat_test)}
}
