% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{ppi_pp_point_and_ci_general}
\alias{ppi_pp_point_and_ci_general}
\title{Tuned PPI++ estimator using numeric lambda search.}
\usage{
ppi_pp_point_and_ci_general(Y_L, f_L, f_U, alpha = 0.1, lambda_range = c(0, 1))
}
\arguments{
\item{Y_L, f_L, f_U}{Observed labels and surrogate scores.}

\item{alpha}{Miscoverage level.}

\item{lambda_range}{Interval to constrain lambda.}
}
\value{
List containing \code{theta}, \code{lambda}, variance, and CI limits.
}
\description{
Tuned PPI++ estimator using numeric lambda search.
}
