% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_estimators.R
\name{ppi_pp_var_hat}
\alias{ppi_pp_var_hat}
\title{Compute PPI++ variance at a given lambda.}
\usage{
ppi_pp_var_hat(lambda, Y_L, f_L, f_U)
}
\arguments{
\item{lambda}{Tuning parameter.}

\item{Y_L, f_L, f_U}{Observed labels and surrogate scores.}
}
\value{
Estimated variance.
}
\description{
Compute PPI++ variance at a given lambda.
}
\keyword{internal}
