% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{train_model_good}
\alias{train_model_good}
\title{Train "good" model: GAM on correct features X1-X5}
\usage{
train_model_good(X, Y)
}
\description{
Knows which features matter, uses flexible functional form
}
\keyword{internal}
