% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{train_model_medium}
\alias{train_model_medium}
\title{Train "medium" model: Random Forest on ALL features}
\usage{
train_model_medium(X, Y)
}
\description{
Uses all 10 features including noise features X6-X10
RF is flexible but will partially fit to noise
}
\keyword{internal}
