% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{train_model_poor_ols}
\alias{train_model_poor_ols}
\title{Train "poor-OLS" model: Linear regression on ALL features}
\usage{
train_model_poor_ols(X, Y)
}
\description{
Uses all 10 features, linear form only
Can't capture nonlinearity + fits to noise features
NOTE: OLS still guarantees E\link{Yhat} = E\link{Y} on training distribution!
}
\keyword{internal}
