% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{train_model_poor_tree}
\alias{train_model_poor_tree}
\title{Train "poor-tree" model: Regression tree on X1 and X6 ONLY}
\usage{
train_model_poor_tree(X, Y)
}
\description{
Uses only X1 (relevant) and X6 (noise) - severely misspecified
KEY: NOT OLS, so no guarantee that \eqn{E[\hat{Y}] = E[Y]} even on training data!
This should show bias even without distribution shift.
}
\keyword{internal}
