% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{eif_linear_discrete_point_and_ci}
\alias{eif_linear_discrete_point_and_ci}
\title{EIF estimator with linear calibration (suboptimal for discrete)}
\usage{
eif_linear_discrete_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
Fits g(Yhat) = a + b*Yhat using linear regression
This is suboptimal when Yhat is discrete but shows importance of proper calibration
}
