% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{fit_calibration_isotonic}
\alias{fit_calibration_isotonic}
\title{Fit calibration model using isotonic regression}
\usage{
fit_calibration_isotonic(Y_cal, Yhat_cal)
}
\description{
Fit calibration model using isotonic regression
}
