% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{fit_calibration_linear}
\alias{fit_calibration_linear}
\title{Fit calibration model using linear regression}
\usage{
fit_calibration_linear(Y_cal, Yhat_cal)
}
\description{
Fit calibration model using linear regression
}
