% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_continuous.R
\name{fit_calibration_spline}
\alias{fit_calibration_spline}
\title{Fit calibration model using cubic spline}
\usage{
fit_calibration_spline(Y_cal, Yhat_cal, df = 4)
}
\description{
Fit calibration model using cubic spline
}
