% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_base_theme_aggregate}
\alias{plot_base_theme_aggregate}
\title{Larger legend theme for aggregate plots (stacked in two rows)}
\usage{
plot_base_theme_aggregate()
}
\description{
Larger legend theme for aggregate plots (stacked in two rows)
}
