% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_discrete.R
\name{ppi_discrete_point_and_ci}
\alias{ppi_discrete_point_and_ci}
\title{PPI estimator for discrete predictions
theta_hat = mean(Yhat_test) + (mean(Y_cal) - mean(Yhat_cal))}
\usage{
ppi_discrete_point_and_ci(Y_cal, Yhat_cal, Yhat_test, alpha = 0.1)
}
\description{
PPI estimator for discrete predictions
theta_hat = mean(Yhat_test) + (mean(Y_cal) - mean(Yhat_cal))
}
